refactor(project): 프로젝트명 변경
- oldboy -> basketclub
