+++ src/main/java/com/basketclub/BasketClubApplication.java
@@ -1,0 +1,11 @@
+package com.basketclub;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BasketClubApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(BasketClubApplication.class, args);
+    }
+}
