+++ src/main/java/com/basketclub/domain/PlayerRepository.java
@@ -1,0 +1,11 @@
+package com.basketclub.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface PlayerRepository extends JpaRepository<Player, Long> {
+    Optional<Player> findByEmail(String email);
+}
