--- src/main/java/oldboy/web/LoginController.java
@@ -1,4 +1,0 @@
-package oldboy.web;
-
-public class LoginController {
-}
