+++ src/test/java/com/basketclub/BasketClubApplicationTest.java
@@ -1,14 +1,14 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package oldboy;
+package com.basketclub;
 
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
 
-class AppTest {
-    @Test void appHasAGreeting() {
-        App classUnderTest = new App();
+class BasketClubApplicationTest {
+    @Test
+    void appHasAGreeting() {
+        BasketClubApplication classUnderTest = new BasketClubApplication();
         assertNotNull(classUnderTest.getGreeting(), "app should have a greeting");
     }
 }
